+++ tests/testthat/test-queue.r
@@ -2,7 +2,7 @@
 test_that('basic', {
   
   # library(jobqueue); library(testthat)
-  q <- expect_silent(Queue$new(workers = 1L, timeout = 5.2))
+  q <- expect_silent(Queue$new(workers = 1L, timeout = 10, wait = FALSE))
   expect_equal(q$state, 'starting')
 
   expect_no_error(q$tmp)
@@ -17,6 +17,7 @@ test_that('basic', {
   job <- expect_silent(q$run({ x + y }, vars = list(x = 3, y = 5)))
   expect_equal(job$result, 8)
 
+  run_now()
   expect_equal(q$state, 'idle')
   expect_equal(length(q$workers), 1L)
 
@@ -26,10 +27,14 @@ test_that('basic', {
   expect_error(q$submit('not a Job'))
   
   expect_silent(q$stop())
-
+  
+  run_now()
   expect_equal(q$state, 'stopped')
   expect_equal(length(q$workers), 0)
   
+  
+  expect_error( Queue$new(init = { stop() }, wait = 10, workers = 1L))
+  expect_error( Queue$new(init = { q()    }, wait = 10, workers = 1L))
 })
 
 
@@ -41,7 +46,8 @@ test_that('config', {
 
   q <- expect_silent(Queue$new(
     workers  = 1L,
-    timeout = 5.3,
+    timeout  = 10,
+    wait     = 10,
     globals  = list(x = 42),
     packages = 'magrittr',
     init     = { y <- 37 },
@@ -52,16 +58,18 @@ test_that('config', {
     )))
 
   q$on('stopped', function () { e$state = 'stopped' })
-  q$on('stopped', class) # A primitive; for code coverage.
+  q$on('stopped', class)() # A primitive; for code coverage.
 
   job <- q$run({ c(x, y) %>% sum })
 
   expect_equal(job$result, 42 + 37)
+  run_now()
   expect_equal(e$state, 'idle')
   expect_equal(e$.next, 'starting')
   expect_equal(e$.star, 'idle')
 
   expect_silent(q$stop())
+  run_now()
   expect_equal(e$state, 'stopped')
   expect_equal(e$.next, 'starting')
   expect_equal(e$.star, 'stopped')
@@ -71,15 +79,17 @@ test_that('config', {
 
 test_that('workers', {
 
-  q <- expect_silent(Queue$new(workers = 2L, max_cpus = 3L, timeout = 5.4))
+  q <- expect_silent(Queue$new(workers = 2L, max_cpus = 3L, wait = FALSE))
 
-  q$run({ Sys.sleep(1) })
-  q$run({ Sys.sleep(1) })
-  q$run({ Sys.sleep(1) })
+  q$run({ Sys.sleep(10) })
+  q$run({ Sys.sleep(10) })
+  q$run({ Sys.sleep(10) })
 
   expect_equal(map(q$jobs, 'state'), rep('queued', 3))
 
-  q$wait('.next')
+  q$wait(state = '.next', timeout = 10)
+  run_now()
+  
   expect_equal(map(q$jobs, 'state'), c('running', 'running', 'queued'))
   expect_equal(map(q$workers, 'state'), rep('busy', 2))
 
@@ -90,16 +100,17 @@ test_that('workers', {
 
 test_that('max_cpus', {
 
-  q <- expect_silent(Queue$new(workers = 3L, max_cpus = 2L, timeout = 5.5))
+  q <- expect_silent(Queue$new(workers = 3L, max_cpus = 2L, wait = FALSE))
 
-  q$run({ Sys.sleep(1) })
-  q$run({ Sys.sleep(1) })
-  q$run({ Sys.sleep(1) })
+  q$run({ Sys.sleep(10) })
+  q$run({ Sys.sleep(10) })
+  q$run({ Sys.sleep(10) })
 
   expect_equal(map(q$jobs, 'state'), rep('queued', 3))
 
-  q$wait('.next')
-  q$run({ Sys.sleep(1) })
+  q$wait(state = 'busy', timeout = 10)
+  q$run({ Sys.sleep(10) })
+  run_now()
 
   expect_equal(map(q$jobs, 'state'), c(rep('running', 2), rep('queued', 2)))
   expect_equal(map(q$workers, 'state'), c('busy', 'busy', 'idle'))
@@ -111,17 +122,17 @@ test_that('max_cpus', {
 
 test_that('interrupt', {
 
-  q <- expect_silent(Queue$new(workers = 1L, timeout = 5.6))
+  q <- expect_silent(Queue$new(workers = 1L, timeout = 10, wait = 10))
 
-  job <- q$run({ Sys.sleep(1) })
+  job <- q$run({ Sys.sleep(10) })
   expect_silent(job$stop())
   expect_s3_class(job$result, class = c('interrupt', 'condition'))
 
-  job <- q$run({ Sys.sleep(1) }, timeout = 0.1)
+  job <- q$run({ Sys.sleep(10) }, timeout = 0.1)
   expect_s3_class(job$result, class = c('interrupt', 'condition'))
 
-  job1 <- q$run({ Sys.sleep(1); 'A' }, stop_id = function (job) 123)
-  job2 <- q$run({ 'B' },               stop_id = ~{ 123 })
+  job1 <- q$run({ Sys.sleep(10); 'A' }, stop_id = function (job) 123)
+  job2 <- q$run({ 'B' },                stop_id = ~{ 123 })
 
   expect_s3_class(job1$result, class = c('interrupt', 'condition'))
   expect_equal(job2$result, 'B')