+++ NEWS.md
@@ -1,3 +1,11 @@
+# jobqueue 1.3.0
+
+* Default to waiting for Queue/Worker to load.
+* Allow functions for more job arguments.
+* Re-export `parallelly::availableCores()`.
+* Disable max_cpu tests on CRAN.
+
+
 # jobqueue 1.0.1
 
 * Initial CRAN submission.